added voyager view model and alert dialog
