+++ gradle.properties
@@ -13,8 +13,8 @@ kotlin.mpp.androidSourceSetLayoutVersion=2
 #Compose
 org.jetbrains.compose.experimental.uikit.enabled=true
 org.jetbrains.compose.experimental.jscanvas.enabled=true
-kotlin.native.cacheKind=none
 
 #Android
 android.useAndroidX=true
 android.nonTransitiveRClass=true
+jvm-target="1.8"
