+++ gradle/libs.versions.toml
@@ -2,7 +2,7 @@
 
 kotlin = "1.9.0"
 agp = "7.4.2"
-compose = "1.4.3"
+compose = "1.5.0"
 androidx-appcompat = "1.6.1"
 androidx-activityCompose = "1.7.2"
 compose-uitooling = "1.4.3"
